refactor: drop legacy learning handlers
